infra/postgres_init/requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
Django==4.2.5
Faker==23.1.0
flake8==6.1.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
